+++ CHANGELOG.md
@@ -14,6 +14,7 @@ Version numbers are based on [Semantic Versioning](https://semver.org).
 
 - Build documentation only for committed files in pre-commit hook
 - Allow multiline texts as localizations in `l10n` XML files
+- Remove `DataUtility` class as data strings are provided by [`LocalizationUtility`](src/classes/Utility/LocalizationUtility.php) class
 - SlackController: Allow adding more data when building message attachment footer
 
 ### Fixed